+++ libs/lib.math/src/main/scala/org/aiotrade/lib/math/timeseries/AbstractTSer.scala
@@ -115,7 +115,7 @@ abstract class AbstractTSer(var freq: TFreq) extends TSer {
     }
   }
 
-  var grids: Array[Float] = Array()
+  var grids: Array[Double] = Array()
   var isOverlapping = false
 
   override def toString: String = {
